requirements-postgres.txt
Flask==3.0.0
Flask-Login==0.6.2
Werkzeug==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.0.0
openpyxl==3.11.0
pytz==2023.3
